split gamma impl from cuda1
